lesson 2

+++ finalProject/Random.h
@@ -1,5 +1,5 @@
 #pragma once
 #ifndef RANDOM
 #define RANDOM
-int getProb(int min,int max);
+float getProb(int min,int max);
 #endif // !RANDOM
